+++ demo/CalendarDemo.jsx
@@ -31,7 +31,7 @@ class Demo extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
-      value: new Date().getTime(),
+      // value: new Date().getTime(),
     };
     this.onSelect = this.onSelect.bind(this);
     this.handleClick = this.handleClick.bind(this);
@@ -81,6 +81,7 @@ class Demo extends React.Component {
       prefixCls: 'kuma-calendar',
       value: moment(me.state.value).locale('zh-cn'),
     };
+    const { value, rangeValue } = this.state;
     return (
       <div className="kuma-form">
         <div
@@ -89,7 +90,9 @@ class Demo extends React.Component {
             width: '400px',
           }}
         >
-          <p>基本</p>
+          <p>
+            基本
+          </p>
           <Calendar
             showToday
             showTime={false}
@@ -99,10 +102,12 @@ class Demo extends React.Component {
             yearSelectOffset={20}
             yearSelectTotal={50}
             size="middle"
-            disabledTime={() => ({
-              disabledMinutes: () => disabledMinutes,
-            })}
-            value={this.state.value}
+            disabledTime={() => (
+              {
+                disabledMinutes: () => disabledMinutes,
+              }
+            )}
+            value={value}
             onSelect={this.onSelect}
             showDateInput
           />
@@ -113,9 +118,11 @@ class Demo extends React.Component {
             width: '400px',
           }}
         >
-          <p>日期格式</p>
+          <p>
+            日期格式
+          </p>
           <Calendar
-            value={this.state.value}
+            value={value}
             format="YYYY/MM/DD"
             onSelect={this.onSelect}
             showDateInput
@@ -127,8 +134,10 @@ class Demo extends React.Component {
             width: '400px',
           }}
         >
-          <p>选择日期</p>
-          <Calendar onSelect={this.onSelect} value={this.state.value} />
+          <p>
+            选择日期
+          </p>
+          <Calendar onSelect={this.onSelect} value={value} />
         </div>
         <div
           className="kuma-form-field"
@@ -136,25 +145,12 @@ class Demo extends React.Component {
             width: '400px',
           }}
         >
-          <p>范围</p>
-          <Calendar disabledDate={disabledDate} value={this.state.value} onSelect={this.onSelect} />
-        </div>
-        <div
-          className="kuma-form-field"
-          style={{
-            width: '400px',
-          }}
-        >
-          <p>时间选择</p>
+          <p>
+            范围
+          </p>
           <Calendar
-            hasTrigger
-            showSecond
-            showHour
-            showTime
-            size="small"
-            timezone={8}
-            locale="en-us"
-            value={this.state.value}
+            disabledDate={disabledDate}
+            value={value}
             onSelect={this.onSelect}
           />
         </div>
@@ -164,8 +160,21 @@ class Demo extends React.Component {
             width: '400px',
           }}
         >
-          <p>禁用</p>
-          <Calendar value={this.state.value} disabled onSelect={this.onSelect} />
+          <p>
+            时间选择
+          </p>
+          <Calendar
+            hasTrigger
+            showSecond
+            showHour
+            showTime
+            size="small"
+            timezone={8}
+            defaultOpenValue={new Date(2018, 8, 1, 3, 12)}
+            locale="en-us"
+            value={value}
+            onSelect={this.onSelect}
+          />
         </div>
         <div
           className="kuma-form-field"
@@ -173,10 +182,23 @@ class Demo extends React.Component {
             width: '400px',
           }}
         >
-          <p>月份</p>
+          <p>
+            禁用
+          </p>
+          <Calendar value={value} disabled onSelect={this.onSelect} />
+        </div>
+        <div
+          className="kuma-form-field"
+          style={{
+            width: '400px',
+          }}
+        >
+          <p>
+            月份
+          </p>
           <MonthCalendar
             size="middle"
-            value={this.state.value}
+            value={value}
             disabledDate={disabledDate}
             onSelect={this.onSelect}
             showDateInput
@@ -188,10 +210,12 @@ class Demo extends React.Component {
             width: '400px',
           }}
         >
-          <p>年份</p>
+          <p>
+            年份
+          </p>
           <YearCalendar
             size="large"
-            value={this.state.value}
+            value={value}
             onSelect={this.onSelect}
             disabledDate={current => { }}
           />
@@ -202,13 +226,14 @@ class Demo extends React.Component {
             width: '400px',
           }}
         >
-          <p>显示日期和日程</p>
           <p>
-            Calendar 通过开放 contentRender 参数来完成日期渲染上的定制，并提供了一个默认的渲染函数
-            Calendar.util.generateContentRender(code) 来完成通用定制。
+            显示日期和日程
+          </p>
+          <p>
+            Calendar 通过开放 contentRender 参数来完成日期渲染上的定制，并提供了一个默认的渲染函数 Calendar.util.generateContentRender(code) 来完成通用定制。
           </p>
           <Calendar
-            value={this.state.value}
+            value={value}
             onSelect={this.onSelect}
             contentRender={Calendar.util.generateContentRender(
               {
@@ -226,20 +251,26 @@ class Demo extends React.Component {
             width: '400px',
           }}
         >
-          <p>直接渲染面板</p>
+          <p>
+            直接渲染面板
+          </p>
           <RcCalendar {...panelOptions} className="panel-demo" />
         </div>
-        <Button onClick={me.handleClick}>changeTime</Button>
+        <Button onClick={me.handleClick}>
+          changeTime
+        </Button>
         <div
           className="kuma-form-field"
           style={{
             width: '400px',
           }}
         >
-          <p>区间日期选择</p>
+          <p>
+            区间日期选择
+          </p>
           <RangeCalendar
             size="large"
-            value={this.state.rangeValue}
+            value={rangeValue}
             onSelect={(v, formatted) => {
               console.log(v, formatted);
               this.onRangeSelect(v, formatted);