+++ src/RcYearCalendar.jsx
@@ -2,10 +2,10 @@ import CalendarMixin from 'rc-calendar/lib/mixin/CalendarMixin';
 import CommonMixin from 'rc-calendar/lib/mixin/CommonMixin';
 // import KeyCode from 'rc-util/lib/KeyCode';
 import React from 'react';
-import CreateClass from 'create-react-class';
+import createClass from 'create-react-class';
 import YearPanel from './YearPanel';
 
-const YearCalendar = CreateClass({
+const YearCalendar = createClass({
   mixins: [CalendarMixin, CommonMixin],
 
   onKeyDown(e) {